+++ ChangeLog.md
@@ -1,4 +1,6 @@
 ## [Pending]
+### Added
+- [List] add filtering
 
 
 ## [0.2.0] - 2024-10-13