+++ ChangeLog.md
@@ -3,6 +3,9 @@
 - [List] add filtering
 - [ModalEndpoint] submit form on Enter
 
+### Fixed
+- [Endpoints] remove endpointStatus when endpoint is deleted
+
 
 ## [0.2.0] - 2024-10-13
 ### Added