+++ ChangeLog.md
@@ -1,4 +1,4 @@
-## [Pending]
+## [0.2.1] - 2024-10-19
 ### Added
 - [List] add filtering
 - [ModalEndpoint] submit form on Enter